third_party/blink/web_tests/platform/generic/storage/indexeddb/objectStore-required-arguments-expected.txt

Test IndexedDB object store required arguments

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

dbname = "objectStore-required-arguments.html"
indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)
objectStore = db.createObjectStore('foo');
PASS objectStore.put(); threw exception TypeError: Failed to execute 'put' on 'IDBObjectStore': 1 argument required, but only 0 present..
PASS objectStore.add(); threw exception TypeError: Failed to execute 'add' on 'IDBObjectStore': 1 argument required, but only 0 present..
PASS objectStore.delete(); threw exception TypeError: Failed to execute 'delete' on 'IDBObjectStore': 1 argument required, but only 0 present..
PASS objectStore.get(); threw exception TypeError: Failed to execute 'get' on 'IDBObjectStore': 1 argument required, but only 0 present..
PASS objectStore.createIndex(); threw exception TypeError: Failed to execute 'createIndex' on 'IDBObjectStore': 2 arguments required, but only 0 present..
PASS objectStore.createIndex('foo'); threw exception TypeError: Failed to execute 'createIndex' on 'IDBObjectStore': 2 arguments required, but only 1 present..
PASS objectStore.index(); threw exception TypeError: Failed to execute 'index' on 'IDBObjectStore': 1 argument required, but only 0 present..
PASS objectStore.deleteIndex(); threw exception TypeError: Failed to execute 'deleteIndex' on 'IDBObjectStore': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
